Working Model (mostly)
I was having a lot of trouble and then I realized: pointers are the
answer. I mean, I've known that since the assignment was assigned, but
for some reason I just wasn't applying what I knew.

This compiles, but not everything works how it should. But it's a good
start...